Move entity caching tests out of crate to `tests/`
