+++ apollo-router/src/plugins/cache/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod cache_control;
-pub(crate) mod entity;
+/// `pub` for tests
+pub mod entity;
 pub(crate) mod invalidation;
 pub(crate) mod invalidation_endpoint;
 pub(crate) mod metrics;
-#[cfg(test)]
 pub(crate) mod tests;
